New version of tests with download
Former-commit-id: d55bbe65e5cd0d732a624429fd716c2ef9bdec14

+++ tests/get_reference.sh
@@ -1,17 +1,31 @@
-curl https://surfdrive.surf.nl/files/index.php/s/e13O5s7PPTsJNli/download -o test.tar -L
-tar -xvf test.tar
-rm test.tar
+if [ -d "tests/tmp" ]; then
+    rm -r "tests/tmp"
+fi
 
-# Copy raw data
-cp -r v1/raw_dir tests
-# Copy transforms data
-for transform in centerline_transform contour_transform heatmap_transform pipeline_transform polar_transform segmentation_transform
-do
-  echo $transform
-  cp -r v1/$transform/* tests/$transform
-done
+if [ ! -d "tests/raw_dir" ]; then
+    curl https://surfdrive.surf.nl/files/index.php/s/e13O5s7PPTsJNli/download -o test.tar -L
+    tar -xvf test.tar
+    rm test.tar
 
-rm -r v1
+    # Copy raw data
+    cp -r v1/raw_dir tests
+    # Copy transforms data
+    for transform in centerline_transform contour_transform heatmap_transform pipeline_transform polar_transform segmentation_transform
+    do
+      if [ -d "tests/$transform/reference" ]; then
+          rm -r tests/$transform/reference
+      fi
+
+      if [ -d "tests/$transform/input" ]; then
+          rm -r tests/$transform/input
+      fi
+      cp -r v1/$transform/* tests/$transform
+    done
+
+    rm -r v1
+else
+    echo "Data was already downloaded. To force a new download remove tests/raw_dir"
+fi
 
 MODELPATH="tests/models"
 
@@ -22,3 +36,4 @@ if [ ! -d $MODELPATH ]; then
     mkdir $MODELPATH/heatmap_transform
     cp models/heatmap_transform/model_0* $MODELPATH/heatmap_transform
     cp -r models/contour_transform_dropout $MODELPATH
+fi
